Split management/public registration file upload
