+++ indico/modules/events/registration/controllers/management/reglists.py
@@ -57,6 +57,7 @@ from indico.modules.events.registration.util import (ActionMenuEntry, create_reg
                                                      import_registrations_from_csv, make_registration_schema)
 from indico.modules.events.registration.views import WPManageRegistration
 from indico.modules.events.util import ZipGeneratorMixin
+from indico.modules.files.controllers import UploadRegistrationFileMixin
 from indico.modules.logs import LogKind
 from indico.modules.logs.util import make_diff_log
 from indico.modules.receipts.models.files import ReceiptFile
@@ -635,6 +636,20 @@ class RHRegistrationTogglePayment(RHManageRegistrationBase):
         if pay != self.registration.is_paid:
             toggle_registration_payment(self.registration, paid=pay)
         return jsonify_data(html=_render_registration_details(self.registration))
+
+
+class RHRegistrationUploadFile(UploadRegistrationFileMixin, RHManageRegFormBase):
+    """
+    Upload a file from a registration form.
+
+    Regform file fields do not wait for the regform to be submitted,
+    but upload the selected files immediately, saving just the generated uuid.
+    Only this uuid is then sent when the regform is submitted.
+    """
+
+    def _process_args(self):
+        RHManageRegFormBase._process_args(self)
+        UploadRegistrationFileMixin._process_args(self)
 
 
 def _modify_registration_status(registration, approve, rejection_reason='', attach_rejection_reason=False):
